Refactor into modular architecture
