+++ index.html
@@ -62,6 +62,6 @@
     <div id="roster-list"></div>
   </section>
 
-  <script src="app.js"></script>
+  <script type="module" src="app.js"></script>
 </body>
 </html>
