+++ modules/design-system.js
@@ -1,0 +1,27 @@
+export const categories = {
+  pts: 'Points',
+  threepm: '3PM',
+  fga: 'FGA',
+  fgm: 'FGM',
+  fta: 'FTA',
+  ftm: 'FTM',
+  reb: 'REB',
+  ast: 'AST',
+  stl: 'STL',
+  blk: 'BLK',
+  tov: 'TOV'
+};
+
+export const defaultWeights = {
+  pts: 1,
+  threepm: 1,
+  fga: -1,
+  fgm: 2,
+  fta: -1,
+  ftm: 1,
+  reb: 1,
+  ast: 2,
+  stl: 4,
+  blk: 4,
+  tov: -2
+};
