+++ modules/eligibility.js
@@ -1,0 +1,3 @@
+export function positionsFor(player) {
+  return player.pos ? [player.pos] : [];
+}
