+++ modules/optimizers.js
@@ -1,0 +1,4 @@
+export function optimizeLineup(players) {
+  // simple placeholder returning players unchanged
+  return players;
+}
