+++ modules/weeks.js
@@ -1,0 +1,3 @@
+export function currentWeek() {
+  return 1;
+}
