+++ modules/workers/calc.worker.js
@@ -1,0 +1,5 @@
+self.onmessage = e => {
+  const { players = [] } = e.data || {};
+  // simple worker echoing back player names
+  postMessage(players.map(p => p.name));
+};
